+++ .changeset/bright-foxes-login-form.md
@@ -1,0 +1,5 @@
+---
+'@shipfox/client-auth': minor
+---
+
+Export the password login form from the continuation entrypoint for downstream route composition.
